// src/types/Commons.type.ts
export type Variant = "primary" | "secondary" | "text" | "error" | "outline";
export type Size = "sm" | "md" | "lg";
export type IconSize = "xs" | Size | "xl";
export type State = "default" | "hover" | "pressed" | "focused" | "disabled";
export type Cursor =
	| "auto"
	| "default"
	| "pointer"
	| "wait"
	| "text"
	| "move"
	| "help"
	| "not-allowed"
	| "crosshair"
	| "grab"
	| "grabbing"
	| "none"
	| "zoom-in"
	| "zoom-out";